# [DLight] Update GEMV rule to support Adreno outer reduction