As Adreno has poor shared memory performance, for LLM workloads, we prefer `q4f16_0` instead of `q4f16_1`.

This PR adds GEMV rule support for Adreno outer reduction, comparing with `q4f16_1`, we get a performance gain

- Prefill(not related with this PR changes): 3636.6501 ms -> 1241.9469 ms
- Decode: 211.0834 ms -> 174.9357 ms